pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "cdpk"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fabdata-llm == 0.27.0",
    "python-dotenv == 1.*",
    "numpy == 2.*",
    "tqdm == 4.*",
    "pandas == 2.2.*",
    "openpyxl == 3.*",
    "matplotlib == 3.*",
    "scikit-learn == 1.5.*",
    "seaborn == 0.13.*",
    "datasets == 3.*",
    "pyfonts == 0.*",
    "pymupdf == 1.24.*",
    "fuzzywuzzy[speedup] == 0.*",
    "azure-identity == 1.*",
    "azure-storage-blob == 12.25.*",
    "dvc >= 3.59.0",
    "dvc-azure >= 3.1.0",
    "ipykernel >= 6.29.5",
    "adjusttext >= 1.3.0",
]

[tool.uv.sources]
fabdata-llm = { git = "https://github.com/AI-for-Education/fabdata-llm.git", rev = "v0.27.0" }

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
    "pdfplumber>=0.11.7",
    "pypdf2>=3.0.1",
]
